xmlsd_generate: encode invalid characters.
<, >, ", and & are invalid in xmlsd attributes or values since this makes xml
impossible to parse. expat already decodes &gt;,&amp;,&lt; and &quot; into their
respecitve characters for us.  So in xmlsd do the reverse of this, encode the
characters into the xml.

Tested with xmlsd_generate on some parsed data and the xml returned was
the same (test will be commited when it has been cleaned up a bit).